% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ASTpct}
\alias{ASTpct}
\title{effective Field Goal Percentage (eFG%) -- eFGpct()}
\usage{
ASTpct(AST, MP, TmMP, TmFG, FG, pct = TRUE, digits = 1)
}
\arguments{
\item{FG}{total number of field goals made (not points)}

\item{pct}{TRUE for percentage of 100, FALSE for fraction of 1}

\item{FGA}{number of field goal attempts}

\item{threeP}{number of 3-points made}
}
\value{
3P%
}
\description{
Calculate effective/adjusted field-goal-percentage (FGP%) using the number of 
sucessfull field goals and total field goal attempts
((FG+0.5*3P)/FGA)*100
}
\details{
assist percentage (AST%) -- ASTpct

Calculate assist percentage - Assist percentage is an 
estimate of the percentage of teammate field goals 
a player assisted while he was on the floor.

the formula is 
AST / (((MP / (Tm MP / 5)) * Tm FG) - FG)
}
