% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_threePperc.R
\name{calc_threePpct}
\alias{calc_threePpct}
\title{3-Point Field Goal Percentage (3P%)}
\usage{
calc_threePpct(threeP, threePA, pct = TRUE)
}
\arguments{
\item{threeP}{total number of succesfull 2-pointers}

\item{threePA}{number of 2-point attempts}

\item{pct}{TRUE for percentage of 100, FALSE for fraction of 1}
}
\value{
3P%
}
\description{
Calculate 3-point-percentage (3P%) using the number of 
sucessfull 3-pointers and total 3-point attempts
(3P/3PA)*100
}
