% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pct3p}
\alias{pct3p}
\title{3-Point Field Goal Percentage (3P%) -- calc3Ppct()}
\usage{
pct3p(p3, p3a, pct = TRUE, digits = 1)
}
\arguments{
\item{p3}{total number of succesfull 2-pointers}

\item{p3a}{number of 2-point attempts}

\item{pct}{TRUE for percentage of 100, FALSE for fraction of 1}

\item{digits}{number of digits}
}
\value{
3P% ( between 0 and 1 )
}
\description{
Calculate 3-point-percentage (3P%) using the number of
sucessfull 3-pointer and total 3-point attempts
(3P/3PA)*100
}
