% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pctFG}
\alias{pctFG}
\title{Field Goal Percentage (FG%) -- pctFG()}
\usage{
pctFG(fg, fga, pct = TRUE, digits = 1)
}
\arguments{
\item{fg}{total number of field goals made (not points)}

\item{fga}{number of field goal attempts}

\item{pct}{TRUE for percentage of 100, FALSE for fraction of 1}
}
\value{
3P%
}
\description{
Calculate field-goal-percentage (FGP%) using the number of 
sucessfull field goals and total field goal attempts
(FG/FGA)*100
}
