% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{playerstats}
\alias{playerstats}
\title{player stats -- playerstats()}
\usage{
playerstats(team = all, player = "all", season = "all",
  competition = "all", type = "rp")
}
\arguments{
\item{team}{team}

\item{player}{player}

\item{season}{season}

\item{competition}{competition}

\item{type}{p for playoffs, r for regular season, rp for both}
}
\value{
table
}
\description{
get player stats for specific season and competition
}
